Created Job Change system

+++ db/migrate/20160713093552_create_jobs.rb
@@ -5,13 +5,13 @@ class CreateJobs < ActiveRecord::Migration
       t.string :description
       t.integer :castle_lvl, :default => 1
       t.integer :lvl, :default => 1
-      t.integer :hp, :default => 0
-      t.integer :mp, :default => 0
-      t.integer :str, :default => 0
-      t.integer :agi, :default => 0
-      t.integer :vit, :default => 0
-      t.integer :int, :default => 0
-      t.integer :luck, :default => 0
+      t.float :hp, :default => 0
+      t.float :mp, :default => 0
+      t.float :str, :default => 0
+      t.float :agi, :default => 0
+      t.float :vit, :default => 0
+      t.float :int, :default => 0
+      t.float :luck, :default => 0
       t.integer :hp_bonus, :default => 0
       t.integer :mp_bonus, :default => 0
       t.integer :str_bonus, :default => 0
